move homeView into a home folder
